requirements.txt
python-telegram-bot==20.3
nest_asyncio==1.5.5
python-dotenv==1.0.0
httpx==0.23.3